Extract declaration formatting out of KotlinInputAstVisitor into DeclarationFormatter
Purely mechanical change that extracts all `visit*` functions related to declarations into corresponding `format*` analogs
